pull demo into containers
